requirements.txt
streamlit
google-generativeai
finance-datareader
yfinance
requests
pandas
beautifulsoup4
lxml